extracted CecilifierContextBasedTestBase from TypeResolverTests to simplify writting unit tests
Basically tests that only requires an 'IVisitorContext' implementation,
i.e, tests that does not require running Cecilifier.
