+++ Cecilifier.Core.Tests/Tests/Unit/ArrayTests.cs
@@ -1,3 +1,4 @@
+using Cecilifier.Core.Tests.Tests.Unit.Framework;
 using Mono.Cecil.Cil;
 using NUnit.Framework;
