+++ Cecilifier.Core.Tests/Tests/Unit/ConstTests.cs
@@ -1,4 +1,5 @@
 using System.Text.RegularExpressions;
+using Cecilifier.Core.Tests.Tests.Unit.Framework;
 using Mono.Cecil.Cil;
 using NUnit.Framework;
