+++ Cecilifier.Core.Tests/Tests/Unit/InterfaceTests.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Text.RegularExpressions;
 using Cecilifier.Core.Extensions;
+using Cecilifier.Core.Tests.Tests.Unit.Framework;
 using NUnit.Framework;
 
 namespace Cecilifier.Core.Tests.Tests.Unit;
