+++ Cecilifier.Core.Tests/Tests/Unit/LocalFunctionTests.cs
@@ -1,4 +1,5 @@
 using System;
+using Cecilifier.Core.Tests.Tests.Unit.Framework;
 using NUnit.Framework;
 
 namespace Cecilifier.Core.Tests.Tests.Unit;
