+++ Cecilifier.Core.Tests/Tests/Unit/Miscellaneous.Assingment.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using Cecilifier.Core.Tests.Tests.Unit.Framework;
 using NUnit.Framework;
 
 namespace Cecilifier.Core.Tests.Tests.Unit
