+++ Cecilifier.Core.Tests/Tests/Unit/MiscellaneousMemberTests.cs
@@ -1,3 +1,4 @@
+using Cecilifier.Core.Tests.Tests.Unit.Framework;
 using NUnit.Framework;
 
 namespace Cecilifier.Core.Tests.Tests.Unit;
