+++ Cecilifier.Core.Tests/Tests/Unit/NameOfTests.cs
@@ -1,3 +1,4 @@
+using Cecilifier.Core.Tests.Tests.Unit.Framework;
 using NUnit.Framework;
 
 namespace Cecilifier.Core.Tests.Tests.Unit
