+++ Cecilifier.Core.Tests/Tests/Unit/OperatorOverloadingTests.cs
@@ -1,3 +1,4 @@
+using Cecilifier.Core.Tests.Tests.Unit.Framework;
 using Microsoft.CodeAnalysis;
 using Mono.Cecil.Cil;
 using static System.Environment;
