+++ Cecilifier.Core.Tests/Tests/Unit/OperatorsTests.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using Cecilifier.Core.Tests.Tests.Unit.Framework;
 using Mono.Cecil.Cil;
 using NUnit.Framework;
