+++ Cecilifier.Core.Tests/Tests/Unit/TypeTests.cs
@@ -1,4 +1,5 @@
 using System.Text.RegularExpressions;
+using Cecilifier.Core.Tests.Tests.Unit.Framework;
 using NUnit.Framework;
 
 namespace Cecilifier.Core.Tests.Tests.Unit;
